+++ packages/ts-newskit/src/components/puzzles/article-sidebar/index.tsx
@@ -68,7 +68,7 @@ export const ArticleSidebar: FC<ArticleSideBarProps> = ({
       />
 
       {data.map(({ title, url, imgUrl }) => (
-        <>
+        <React.Fragment key={title}>
           <CardComposable
             columns="0fr 1fr"
             overrides={{
@@ -104,7 +104,7 @@ export const ArticleSidebar: FC<ArticleSideBarProps> = ({
               stylePreset: 'dashedDivider'
             }}
           />
-        </>
+        </React.Fragment>
       ))}
     </Block>
   );