+++ packages/ts-newskit/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.146.1](https://github.com/newsuk/times-components/compare/@times-components/ts-newskit@1.146.0...@times-components/ts-newskit@1.146.1) (2023-12-08)
+
+**Note:** Version bump only for package @times-components/ts-newskit
+
+
+
+
+
 # [1.146.0](https://github.com/newsuk/times-components/compare/@times-components/ts-newskit@1.145.0...@times-components/ts-newskit@1.146.0) (2023-12-07)